+++ lib/log_manager_plugin.dart
@@ -1,5 +1,5 @@
 export 'src/log_config.dart';
-export 'src/log_util.dart';
+export 'src/log_manager.dart';
 export 'src/log_file_manager.dart';
 export 'src/dio_log_interceptor.dart';
 export 'src/log_viewer_page.dart';